requirements.txt
pyqtgraph
pyqt5
scipy
netifaces
cdp-py
PyOpenGL
PyOpenGL_accelerate